backend/requirements.txt
flask
flask_cors
mysql_connector
configparser
simplejson
sqlalchemy
